refreshToken

+++ src/routes/article.ts
@@ -1,6 +1,7 @@
 import { Request, Response, Router } from 'express';
 import { asyncWrap } from '../utils/asyncWrapper';
 import database from '../config/database';
+import verifyToken from '../middleware/verifyToken';
 
 const router = Router();
 
@@ -18,7 +19,6 @@ router.get(
 			const [row] = await database.query(
 				`SELECT * FROM POSTS WHERE ID = ${id};`
 			);
-			console.log(row);
 			res.send(row);
 		} catch (e) {
 			console.log(e);
@@ -28,11 +28,12 @@ router.get(
 );
 router.post(
 	'/:id/edit',
+	verifyToken,
 	asyncWrap(async (req: Request<ArticleProps>, res) => {
-		const { user_id, title, content, preview_text } = req.body;
+		const { title, content, preview_text } = req.body;
 		try {
 			const data = await database.query(
-				`UPDATE POSTS SET content = '${content}', preview_text='${preview_text}', updated_at = NOW(),  title='${title}' WHERE id ='${req.params.id}'`
+				`UPDATE POSTS SET content = '${content}', preview_text='${preview_text}', updated_at = NOW(),  title='${title}' WHERE id ='${req.params.id}' and user_id= '${req.user}'`
 			);
 			res.send(data);
 		} catch (e) {
@@ -43,10 +44,12 @@ router.post(
 );
 router.delete(
 	'/:id',
+	verifyToken,
 	asyncWrap(async (req: Request<ArticleProps>, res) => {
+		const id = req.user;
 		try {
 			const data = await database.query(
-				`DELETE FROM POSTS WHERE id ='${req.params.id}'`
+				`DELETE FROM POSTS WHERE id ='${req.params.id}' AND user_id = '${id}'`
 			);
 			res.send(data);
 		} catch (e) {
